requirements.txt
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2025.4.26
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.11.0
einops==0.8.1
fonttools==4.55.3
fsspec==2025.3.2
gmpy2==2.2.1
huggingface-hub==0.30.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.6
kiwisolver==1.4.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numexpr==2.10.1
numpy==1.23.5
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
ply==3.11
psutil==5.9.0
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.13.1
sip==6.7.12
six==1.17.0
sympy==1.13.3
timm==0.9.12
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
tzdata==2025.2
urllib3==2.3.0
zipp==3.21.0
